修复编译错误并恢复权限问题排查文档
修复内容：
1. 修复Program.cs中的类型转换错误
   - 将long类型的maxFileSize显式转换为int类型
   - 解决ValueLengthLimit赋值的类型不匹配问题

2. 修复StudentMockExamController中的依赖注入问题
   - 添加ApplicationDbContext依赖注入
   - 添加必要的using语句（EntityFrameworkCore）
   - 修复诊断方法中_context未定义的错误

3. 恢复权限问题排查文档
   - 重新创建MockExamPermissionTroubleshooting.md文档
   - 包含完整的权限问题诊断和修复指南
   - 提供详细的API使用示例和排查步骤

所有编译错误已解决，项目可以正常构建。

+++ ExaminaWebApplication/Program.cs
@@ -72,7 +72,7 @@ builder.Services.Configure<Microsoft.AspNetCore.Http.Features.FormOptions>(optio
     options.MultipartBodyLengthLimit = maxFileSize;
 
     // 设置单个文件大小限制
-    options.ValueLengthLimit = maxFileSize;
+    options.ValueLengthLimit = (int)maxFileSize;
 
     // 设置表单键数量限制
     options.KeyLengthLimit = 2048;
